feat(Tooltip): refactor NewTooltip component and update imports
