+++ libs/ui/src/index.ts
@@ -52,6 +52,7 @@ export * from './pure/TokenAmount'
 export * from './pure/TokenName'
 export * from './pure/TokenSymbol'
 export * from './pure/Tooltip'
+export * from './pure/Tooltip/Tooltip'
 export * from './pure/TruncatedText'
 export * from './styles'
 export * from './theme'
