+++ libs/ui/src/pure/Tooltip/index.tsx
@@ -45,6 +45,16 @@ export interface HoverTooltipProps extends Omit<PopoverProps, 'content' | 'show'
   tooltipCloseDelay?: number
 }
 
+/**
+ * Tooltip that appears when hovering over the children
+ *
+ * @see HelpTooltip as an alternative if you need to show a tooltip with a question mark icon (or icon of your choice)
+ * @see InfoTooltip as an alternative if you need to show a tooltip with an info icon
+ * @see Tooltip as an alternative if you need to control when the tooltip is shown
+ *
+ * @param props
+ * @returns
+ */
 // TODO: Break down this large function into smaller functions
 // TODO: Add proper return type annotation
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
@@ -224,5 +234,3 @@ export function renderTooltip(tooltip: ReactNode | ((params?: any) => ReactNode)
   }
   return tooltip
 }
-
-export * from './NewTooltip'
